backend/requirements.txt
# ============================================================
# Previa App Backend — Python Dependencies
# ============================================================

# --- Web Framework ------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# --- Database -----------------------------------------------
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1
asyncpg

# --- Data Processing ----------------------------------------
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# --- HTTP Client --------------------------------------------
httpx==0.26.0

# --- LLM & Agent Framework ----------------------------------
langchain==0.1.6
langchain-anthropic==0.1.4
anthropic==0.18.1

# --- Authentication & Security ------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# --- Rate Limiting ------------------------------------------
slowapi==0.1.9

# --- Web Scraping & ingestion (DOF, SAT Datos Abiertos) ------
playwright==1.41.0
beautifulsoup4==4.12.3

# --- Payments -----------------------------------------------
stripe

# --- Monitoring & Error Tracking ----------------------------
sentry-sdk[fastapi]

# --- Utilities ----------------------------------------------
python-dotenv==1.0.1
pydantic[email]==2.5.3
pydantic-settings>=2.1.0
pypdf
apscheduler>=3.10.0

# --- Development & Testing ----------------------------------
pytest==8.0.0
pytest-asyncio==0.23.3
